+++ code/src/sixsq/nuvla/auth/cookies.clj
@@ -50,12 +50,17 @@
 
 
 (defn collect-groups-for-user
+  [user-id]
+  (-> (crud/query-as-admin
+        group/resource-type
+        {:cimi-params {:filter (parser/parse-cimi-filter (format "users='%s'" user-id))
+                       :select ["id"]}})
+      second))
+
+
+(defn collect-groups-set-for-user
   [id]
-  (->> (crud/query-as-admin
-         group/resource-type
-         {:cimi-params {:filter (parser/parse-cimi-filter (format "users='%s'" id))
-                        :select ["id"]}})
-       second
+  (->> (collect-groups-for-user id)
        (map :id)
        set))
 
@@ -63,7 +68,7 @@
 (defn create-cookie-info
   [user-id & {:keys [session-id headers client-ip active-claim claims roles-ext]}]
   (let [server               (:nuvla-ssl-server-name headers)
-        collected-groups-set (collect-groups-for-user user-id)
+        collected-groups-set (collect-groups-set-for-user user-id)
         groups               (-> collected-groups-set
                                  sort
                                  seq)]