+++ code/src/sixsq/nuvla/server/resources/callback_join_group.clj
@@ -13,8 +13,8 @@ visited, the email identifier is marked as validated.
     [sixsq.nuvla.server.resources.common.std-crud :as std-crud]
     [sixsq.nuvla.server.resources.common.utils :as u]
     [sixsq.nuvla.server.resources.credential-hashed-password :as hashed-password]
-    [sixsq.nuvla.server.resources.user-template-minimum :as user-template-minimum]
     [sixsq.nuvla.server.resources.user-template :as user-template]
+    [sixsq.nuvla.server.resources.user-template-minimum :as user-template-minimum]
     [sixsq.nuvla.server.resources.user.utils :as user-utils]
     [sixsq.nuvla.server.util.response :as r]))
 
@@ -32,7 +32,7 @@ visited, the email identifier is marked as validated.
   (std-crud/add-if-absent
     (str user-utils/resource-url " '" email "'") user-utils/resource-url
     {:template
-     {:href     (str user-template/resource-type "/" user-template-minimum/resource-url)
+     {:href     (str user-template/resource-type "/" user-template-minimum/registration-method)
       :email    email
       :password new-password}})
   (auth-password/identifier->user-id email))